replace _OperatorLazyWrapper with mixins
